Rename and move ReaderItemDecoration for broader reuse
